move remaining constants to new taxonomy.tsx file
