+++ frontend/src/scenes/session-recordings/player/inspector/components/eventPropertyFilteringLogic.ts
@@ -1,6 +1,6 @@
 import { connect, kea, path, selectors } from 'kea'
 import { userPreferencesLogic } from 'lib/logic/userPreferencesLogic'
-import { CLOUD_INTERNAL_POSTHOG_PROPERTY_KEYS } from 'lib/taxonomy'
+import { CLOUD_INTERNAL_POSTHOG_PROPERTY_KEYS } from '~/taxonomy/taxonomy'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 
 import { PROPERTY_KEYS } from '~/taxonomy/taxonomy'
